Moved SchemeResults to output folder
